requirements.txt
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Django==2.1.5
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
Markdown==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
